+++ CHANGES.rst
@@ -14,6 +14,7 @@ CHANGELOG
 * POI list are not sorted by category/alphabetic, API order (progression) is kept (fixes #56)
 * Add departure city in results list (fixes #57)
 * Increase result thumbnail size to match height (ref #57)
+* Switch parking popup to label (ref #55)
 
 1.24 (2013-08-27)
 -----------------